+++ example/27_layernorm/elementwise_layernorm_blockwise.cpp
@@ -9,7 +9,7 @@
 
 #include "ck/ck.hpp"
 #include "ck/utility/reduction_enums.hpp"
-#include "ck/tensor_operation/gpu/device/device_elementwise_layernorm_impl.hpp"
+#include "ck/tensor_operation/gpu/device/impl/device_elementwise_layernorm_impl.hpp"
 #include "ck/tensor_operation/gpu/device/reduction_operator_mapping.hpp"
 
 #include "ck/library/utility/check_err.hpp"
